Avoid interactive prompt with easyrsa tool
On Ubuntu 18.04, generating certs and keys was failing
because `easyrsa` would prompt for user input. This is of
course undesirable in the context of Ansible, so adding the
`--batch` option causes it to just use the defaults without
prompting.

Also, batch mode seems to ignore the set CN, so we must set
it explicitly.